PSFR-1220 Fix check your answers on fresh session
* Show submitted input if session is changed
* Stop gotenberg port in the docker-compose-test file clashing with the
  port of npm run start:dev
