+++ feature.env
@@ -12,4 +12,4 @@ MANAGE_USERS_API_URL=http://localhost:9091/manageUsersApi
 RESETTLEMENT_PASSPORT_API_URL=http://localhost:9091/rpApi
 FEATURE_FLAG_ENABLED=false
 LOCAL_FEATURE_FLAG_ENABLED=true
-GOTENBERG_API_URL=http://localhost:3001
+GOTENBERG_API_URL=http://localhost:3009
