+++ server/data/assessmentStateService.test.ts
@@ -1,7 +1,7 @@
 import { AssessmentStateService, StateKey } from './assessmentStateService'
 import AssessmentStore from './assessmentStore'
 import { createRedisClient } from './redisClient'
-import { AssessmentPage, SubmittedInput } from './model/BCST2Form'
+import { AssessmentPage, QuestionsAndAnswers, SubmittedInput, SubmittedQuestionAndAnswer } from './model/BCST2Form'
 
 jest.mock('./assessmentStore')
 
@@ -467,4 +467,46 @@ describe('assessmentStateService', () => {
       expect(deleteQuestionListSpy).toHaveBeenCalledTimes(0)
     })
   })
+
+  describe('mergeQuestionsAndAnswers', () => {
+    it('Should merge questions and answers favouring questions from the cache', () => {
+      const apiQAndA: AssessmentPage = {
+        questionsAndAnswers: [aQuestionAndAnswer('1', 'API 1'), aQuestionAndAnswer('2', 'API 2')],
+      } as AssessmentPage
+
+      const cacheQAndA: SubmittedInput = {
+        questionsAndAnswers: [aSubmittedQAndA('1', 'Cache 1'), aSubmittedQAndA('3', 'Cache 3')],
+      } as SubmittedInput
+
+      const merged = assessmentStateService.mergeQuestionsAndAnswers(apiQAndA, cacheQAndA)
+
+      expect(merged).toHaveLength(3)
+      const mergedItems = merged.map(i => {
+        return { q: i.question, a: i.answer.answer }
+      })
+      expect(mergedItems).toContainEqual({ q: '1', a: 'Cache 1' })
+      expect(mergedItems).toContainEqual({ q: '2', a: 'API 2' })
+      expect(mergedItems).toContainEqual({ q: '3', a: 'Cache 3' })
+    })
+  })
 })
+
+function aQuestionAndAnswer(id: string, answer: string): QuestionsAndAnswers {
+  return {
+    question: {
+      id,
+    },
+    answer: {
+      answer,
+    },
+  } as QuestionsAndAnswers
+}
+
+function aSubmittedQAndA(id: string, answer: string): SubmittedQuestionAndAnswer {
+  return {
+    question: id,
+    answer: {
+      answer,
+    },
+  } as SubmittedQuestionAndAnswer
+}
